refactor: apply SRP to Frontend — extract hook, move modal, remove dead API
- IncidentForm now uses useIncidentForm hook (state + API logic extracted)
- Move SuccessModal from incidents/ to modal/ (reusable component)
- Remove dead getComplaint method from complaintsService (endpoint removed)
- Remove dead getComplaint tests from service test file
- Component is now pure presentation + form action wrapper

Resolves: AUDITORIA §3.1 (Frontend SRP — extract hooks, separate modal)
